fixed for -Wconversion

+++ include/votca/tools/cubicspline.h
@@ -98,7 +98,7 @@ class CubicSpline : public Spline {
    * one entry in that fitting matrix.
    */
   template <typename matrix_type>
-  void AddToFitMatrix(matrix_type &A, double x, int offset1, int offset2 = 0,
+  void AddToFitMatrix(matrix_type &A, double x, long int offset1, long int offset2 = 0,
                       double scale = 1);
 
   /**
@@ -112,7 +112,7 @@ class CubicSpline : public Spline {
    * one entry in that fitting matrix.
    */
   template <typename matrix_type>
-  void AddToFitMatrix(matrix_type &A, double x, int offset1, int offset2,
+  void AddToFitMatrix(matrix_type &A, double x, int long offset1, int long offset2,
                       double scale1, double scale2);
 
   /**
@@ -123,8 +123,8 @@ class CubicSpline : public Spline {
    * Same as previous function, but vector-valued and with scale=1.0
    */
   template <typename matrix_type, typename vector_type>
-  void AddToFitMatrix(matrix_type &M, vector_type &x, int offset1,
-                      int offset2 = 0);
+  void AddToFitMatrix(matrix_type &M, vector_type &x, long int offset1,
+                      long int offset2 = 0);
 
   /**
    * \brief Add boundary condition of sum_i f_i =0 to fitting matrix
@@ -132,7 +132,7 @@ class CubicSpline : public Spline {
    * \param offsets
    */
   template <typename matrix_type>
-  void AddBCSumZeroToFitMatrix(matrix_type &A, int offset1, int offset2 = 0);
+  void AddBCSumZeroToFitMatrix(matrix_type &A, long int offset1, long int offset2 = 0);
 
   /**
    * \brief Add boundary conditions to fitting matrix
@@ -140,7 +140,7 @@ class CubicSpline : public Spline {
    * \param offsets
    */
   template <typename matrix_type>
-  void AddBCToFitMatrix(matrix_type &A, int offset1, int offset2 = 0);
+  void AddBCToFitMatrix(matrix_type &A, long int offset1, long int offset2 = 0);
 
  protected:
   // A spline can be written in the form
@@ -180,8 +180,8 @@ inline double CubicSpline::CalculateDerivative(const double &r) {
 }
 
 template <typename matrix_type>
-inline void CubicSpline::AddToFitMatrix(matrix_type &M, double x, int offset1,
-                                        int offset2, double scale) {
+inline void CubicSpline::AddToFitMatrix(matrix_type &M, double x, long int offset1,
+                                        long int offset2, double scale) {
   long int spi = getInterval(x);
   M(offset1, offset2 + spi) += A(x) * scale;
   M(offset1, offset2 + spi + 1) += B(x) * scale;
@@ -191,8 +191,8 @@ inline void CubicSpline::AddToFitMatrix(matrix_type &M, double x, int offset1,
 
 // for adding f'(x)*scale1 + f(x)*scale2 as needed for threebody interactions
 template <typename matrix_type>
-inline void CubicSpline::AddToFitMatrix(matrix_type &M, double x, int offset1,
-                                        int offset2, double scale1,
+inline void CubicSpline::AddToFitMatrix(matrix_type &M, double x, long int offset1,
+                                        long int offset2, double scale1,
                                         double scale2) {
   long int spi = getInterval(x);
   M(offset1, offset2 + spi) += Aprime(x) * scale1;
@@ -208,7 +208,7 @@ inline void CubicSpline::AddToFitMatrix(matrix_type &M, double x, int offset1,
 
 template <typename matrix_type, typename vector_type>
 inline void CubicSpline::AddToFitMatrix(matrix_type &M, vector_type &x,
-                                        int offset1, int offset2) {
+                                        long int offset1, long int offset2) {
   for (int i = 0; i < x.size(); ++i) {
     long int spi = getInterval(x(i));
     M(offset1 + i, offset2 + spi) = A(x(i));
@@ -219,8 +219,8 @@ inline void CubicSpline::AddToFitMatrix(matrix_type &M, vector_type &x,
 }
 
 template <typename matrix_type>
-inline void CubicSpline::AddBCSumZeroToFitMatrix(matrix_type &M, int offset1,
-                                                 int offset2) {
+inline void CubicSpline::AddBCSumZeroToFitMatrix(matrix_type &M, long int offset1,
+                                                 long int offset2) {
   for (int i = 0; i < _r.size(); ++i) {
     M(offset1, offset2 + i) = 1;
     M(offset1, offset2 + _r.size() + i) = 0;
@@ -228,8 +228,8 @@ inline void CubicSpline::AddBCSumZeroToFitMatrix(matrix_type &M, int offset1,
 }
 
 template <typename matrix_type>
-inline void CubicSpline::AddBCToFitMatrix(matrix_type &M, int offset1,
-                                          int offset2) {
+inline void CubicSpline::AddBCToFitMatrix(matrix_type &M, long int offset1,
+                                          long int offset2) {
   for (int i = 0; i < _r.size() - 2; ++i) {
     M(offset1 + i + 1, offset2 + i) = A_prime_l(i);
     M(offset1 + i + 1, offset2 + i + 1) = B_prime_l(i) - A_prime_r(i);
