+++ include/votca/tools/table.h
@@ -40,24 +40,24 @@ class Table {
   void resize(long int N);
   long int size() const { return _x.size(); }
 
-  double &x(int i) { return _x[i]; }
-  double &y(int i) { return _y[i]; }
+  double &x(long int i) { return _x[i]; }
+  double &y(long int i) { return _y[i]; }
 
-  const double &x(int i) const { return _x[i]; }
-  const double &y(int i) const { return _y[i]; }
-  char &flags(int i) { return _flags[i]; }
-  double &yerr(int i) { return _yerr[i]; }
+  const double &x(long int i) const { return _x[i]; }
+  const double &y(long int i) const { return _y[i]; }
+  char &flags(long int i) { return _flags[i]; }
+  double &yerr(long int i) { return _yerr[i]; }
 
-  void set(const int &i, const double &x, const double &y) {
+  void set(const long int &i, const double &x, const double &y) {
     _x[i] = x;
     _y[i] = y;
   }
-  void set(const int &i, const double &x, const double &y, const char &flags) {
+  void set(const long int &i, const double &x, const double &y, const char &flags) {
     _x[i] = x;
     _y[i] = y;
     _flags[i] = flags;
   }
-  void set(const int &i, const double &x, const double &y, const char &flags,
+  void set(const long int &i, const double &x, const double &y, const char &flags,
            const double &yerr) {
     _x[i] = x;
     _y[i] = y;
@@ -73,7 +73,7 @@ class Table {
   void Load(std::string filename);
   void Save(std::string filename) const;
 
-  void Smooth(int Nsmooth);
+  void Smooth(long int Nsmooth);
 
   bool GetHasYErr() { return _has_yerr; }
   void SetHasYErr(bool has_yerr) { _has_yerr = has_yerr; }
@@ -131,7 +131,7 @@ inline std::ostream &operator<<(std::ostream &out, const Table &t) {
   out.precision(10);
 
   if (t._has_yerr) {
-    for (int i = 0; i < t._x.size(); ++i) {
+    for (long int i = 0; i < t._x.size(); ++i) {
       out << t._x[i] << " " << t._y[i] << " " << t._yerr[i];
       if (t._flags[i] == '\0' || t._flags[i] == ' ') {
         out << "\n";
@@ -140,7 +140,7 @@ inline std::ostream &operator<<(std::ostream &out, const Table &t) {
       }
     }
   } else {
-    for (int i = 0; i < t._x.size(); ++i) {
+    for (long int i = 0; i < t._x.size(); ++i) {
       out << t._x[i] << " " << t._y[i];
       if (t._flags[i] == '\0' || t._flags[i] == ' ') {
         out << "\n";
